create FS component, add useEffect
