+++ packages/client/src/pages/GamePage/Components/gameBlock.module.scss
@@ -37,3 +37,8 @@
   width: 50px;
   height: 50px;
 }
+
+.fullscreen-icon-container {
+  position: absolute;
+  z-index: 10;
+}
